+++ src/Compilers/CSharp/Portable/Symbols/Synthesized/Records/SynthesizedRecordConstructor.cs
@@ -3,14 +3,10 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Diagnostics;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading;
-using Microsoft.CodeAnalysis.CSharp.Symbols;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using Microsoft.CodeAnalysis.PooledObjects;
 using Roslyn.Utilities;
 
 namespace Microsoft.CodeAnalysis.CSharp.Symbols
@@ -101,743 +97,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Symbols
 
             Interlocked.CompareExchange(ref _capturedParameters, Binder.CapturedParametersFinder.GetCapturedParameters(this), null);
             return _capturedParameters;
-        }
-    }
-}
-
-namespace Microsoft.CodeAnalysis.CSharp
-{
-    // PROTOTYPE(PrimaryConstructors): Move to a Binder specific file?
-    partial class Binder
-    {
-        internal sealed class CapturedParametersFinder : IdentifierUsedAsValueFinder
-        {
-            private readonly SynthesizedPrimaryConstructor _primaryConstructor;
-            private readonly HashSet<string> _namesToCheck;
-            private readonly ArrayBuilder<ParameterSymbol> _captured;
-
-            private CapturedParametersFinder(SynthesizedPrimaryConstructor primaryConstructor, HashSet<string> namesToCheck, ArrayBuilder<ParameterSymbol> captured)
-            {
-                this._primaryConstructor = primaryConstructor;
-                this._namesToCheck = namesToCheck;
-                this._captured = captured;
-            }
-
-            public static IReadOnlyDictionary<ParameterSymbol, FieldSymbol> GetCapturedParameters(SynthesizedPrimaryConstructor primaryConstructor)
-            {
-                var namesToCheck = PooledHashSet<string>.GetInstance();
-                addParameterNames(namesToCheck);
-
-                if (namesToCheck.Count == 0)
-                {
-                    namesToCheck.Free();
-                    return SpecializedCollections.EmptyReadOnlyDictionary<ParameterSymbol, FieldSymbol>();
-                }
-
-                var captured = ArrayBuilder<ParameterSymbol>.GetInstance(primaryConstructor.Parameters.Length);
-
-                var finder = new CapturedParametersFinder(primaryConstructor, namesToCheck, captured);
-
-                var containingType = primaryConstructor.ContainingType;
-
-                foreach (var member in containingType.GetMembers())
-                {
-                    Binder? bodyBinder;
-                    CSharpSyntaxNode? syntaxNode;
-
-                    getBodyBinderAndSyntaxIfPossiblyCapturingMethod(member, out bodyBinder, out syntaxNode);
-                    if (bodyBinder is null)
-                    {
-                        continue;
-                    }
-
-                    Debug.Assert(syntaxNode is not null);
-
-                    bool keepChecking = checkParameterReferencesInMethodBody(syntaxNode, bodyBinder);
-                    if (!keepChecking)
-                    {
-                        break;
-                    }
-                }
-
-                finder.Free();
-                namesToCheck.Free();
-
-                if (captured.Count == 0)
-                {
-                    captured.Free();
-                    return SpecializedCollections.EmptyReadOnlyDictionary<ParameterSymbol, FieldSymbol>();
-                }
-
-                var result = new Dictionary<ParameterSymbol, FieldSymbol>(ReferenceEqualityComparer.Instance);
-
-                foreach (var parameter in captured)
-                {
-                    // PROTOTYPE(PrimaryConstructors): Figure out naming strategy
-                    string name = "<" + parameter.Name + ">PC__BackingField";
-
-                    // PROTOTYPE(PrimaryConstructors): Ever read-only?
-                    result.Add(parameter, new SynthesizedFieldSymbol(containingType, parameter.Type, name));
-                }
-
-                captured.Free();
-                return result;
-
-                void addParameterNames(PooledHashSet<string> namesToCheck)
-                {
-                    foreach (var parameter in primaryConstructor.Parameters)
-                    {
-                        if (parameter.Name.Length != 0)
-                        {
-                            namesToCheck.Add(parameter.Name);
-                        }
-                    }
-                }
-
-                void getBodyBinderAndSyntaxIfPossiblyCapturingMethod(Symbol member, out Binder? bodyBinder, out CSharpSyntaxNode? syntaxNode)
-                {
-                    bodyBinder = null;
-                    syntaxNode = null;
-
-                    if ((object)member == primaryConstructor)
-                    {
-                        return;
-                    }
-
-                    if (member.IsStatic ||
-                        !(member is MethodSymbol method && MethodCompiler.GetMethodToCompile(method) is SourceMemberMethodSymbol sourceMethod))
-                    {
-                        return;
-                    }
-
-                    if (sourceMethod.IsExtern)
-                    {
-                        return;
-                    }
-
-                    bodyBinder = sourceMethod.TryGetBodyBinder();
-
-                    if (bodyBinder is null)
-                    {
-                        return;
-                    }
-
-                    syntaxNode = sourceMethod.SyntaxNode;
-                }
-
-                bool checkParameterReferencesInMethodBody(CSharpSyntaxNode syntaxNode, Binder bodyBinder)
-                {
-                    switch (syntaxNode)
-                    {
-                        case ConstructorDeclarationSyntax s:
-                            return finder.CheckIdentifiersInNode(s.Initializer, bodyBinder) &&
-                                   finder.CheckIdentifiersInNode(s.Body, bodyBinder) &&
-                                   finder.CheckIdentifiersInNode(s.ExpressionBody, bodyBinder);
-
-                        case BaseMethodDeclarationSyntax s:
-                            return finder.CheckIdentifiersInNode(s.Body, bodyBinder) &&
-                                   finder.CheckIdentifiersInNode(s.ExpressionBody, bodyBinder);
-
-                        case AccessorDeclarationSyntax s:
-                            return finder.CheckIdentifiersInNode(s.Body, bodyBinder) &&
-                                   finder.CheckIdentifiersInNode(s.ExpressionBody, bodyBinder);
-
-                        case ArrowExpressionClauseSyntax s:
-                            return finder.CheckIdentifiersInNode(s, bodyBinder);
-
-                        default:
-                            throw ExceptionUtilities.UnexpectedValue(syntaxNode);
-                    }
-                }
-            }
-
-            protected override bool IsIdentifierOfInterest(IdentifierNameSyntax id)
-            {
-                return _namesToCheck.Contains(id.Identifier.ValueText);
-            }
-
-            protected override bool CheckAndClearLookupResult(Binder enclosingBinder, IdentifierNameSyntax id, LookupResult lookupResult)
-            {
-                if (lookupResult.IsMultiViable)
-                {
-                    bool? isInsideNameof = null;
-                    bool detectedCapture = false;
-
-                    foreach (var candidate in lookupResult.Symbols)
-                    {
-                        if (candidate is ParameterSymbol parameter && parameter.ContainingSymbol == (object)_primaryConstructor)
-                        {
-                            isInsideNameof ??= enclosingBinder.IsInsideNameof;
-
-                            if (isInsideNameof.GetValueOrDefault())
-                            {
-                                break;
-                            }
-                            else if (lookupResult.IsSingleViable)
-                            {
-                                Debug.Assert(lookupResult.SingleSymbolOrDefault == (object)parameter);
-
-                                // Check for left of potential color color member access 
-                                if (isTypeOrValueReceiver(enclosingBinder, id, parameter.Type, out SyntaxNode? memberAccessNode, out string? memberName, out int targetMemberArity, out bool invoked))
-                                {
-                                    lookupResult.Clear();
-                                    if (TreatAsInstanceMemberAccess(enclosingBinder, parameter.Type, memberAccessNode, memberName, targetMemberArity, invoked, lookupResult))
-                                    {
-                                        _captured.Add(parameter);
-                                        detectedCapture = true;
-                                    }
-
-                                    // We cleared the lookupResult and the candidate list within it.
-                                    // Do not attempt to continue the enclosing foreach
-                                    break;
-                                }
-                            }
-
-                            _captured.Add(parameter);
-                            detectedCapture = true;
-                        }
-                    }
-
-                    if (detectedCapture)
-                    {
-                        _namesToCheck.Remove(id.Identifier.ValueText);
-
-                        if (_namesToCheck.Count == 0)
-                        {
-                            return false;
-                        }
-                    }
-                }
-
-                lookupResult.Clear();
-                return true;
-            }
-        }
-
-        internal abstract class IdentifierUsedAsValueFinder
-        {
-            private LookupResult? _lookupResult;
-
-            protected void Free()
-            {
-                _lookupResult?.Free();
-            }
-
-            protected bool CheckIdentifiersInNode(CSharpSyntaxNode? node, Binder binder)
-            {
-                if (node == null)
-                {
-                    return true;
-                }
-
-                var nodesOfInterest = node.DescendantNodesAndSelf(descendIntoChildren: childrenNeedChecking, descendIntoTrivia: false);
-
-                foreach (var n in nodesOfInterest)
-                {
-                    Binder enclosingBinder = getEnclosingBinderForNode(contextNode: node, contextBinder: binder, n);
-
-                    switch (n)
-                    {
-                        case AnonymousFunctionExpressionSyntax lambdaSyntax:
-                            if (!CheckLambda(lambdaSyntax, enclosingBinder))
-                            {
-                                return false;
-                            }
-
-                            break;
-
-                        case IdentifierNameSyntax id:
-
-                            switch (id.Parent)
-                            {
-                                case MemberAccessExpressionSyntax memberAccess:
-                                    if (memberAccess.Expression != id)
-                                    {
-                                        continue;
-                                    }
-                                    break;
-
-                                case QualifiedNameSyntax qualifiedName:
-                                    if (qualifiedName.Left != id)
-                                    {
-                                        continue;
-                                    }
-                                    break;
-
-                                case AssignmentExpressionSyntax assignment:
-                                    if (assignment.Left == id &&
-                                        assignment.Parent?.Kind() is SyntaxKind.ObjectInitializerExpression or SyntaxKind.WithInitializerExpression)
-                                    {
-                                        continue;
-                                    }
-                                    break;
-                            }
-
-                            if (SyntaxFacts.IsInTypeOnlyContext(id) &&
-                                !(id.Parent is BinaryExpressionSyntax { RawKind: (int)SyntaxKind.IsExpression } isExpression &&
-                                    isExpression.Right == id))
-                            {
-                                continue;
-                            }
-
-                            if (!IsIdentifierOfInterest(id))
-                            {
-                                continue;
-                            }
-
-                            if (!CheckIdentifier(enclosingBinder, id))
-                            {
-                                return false;
-                            }
-
-                            break;
-
-                        case QueryExpressionSyntax query:
-                            if (!CheckQuery(query, enclosingBinder))
-                            {
-                                return false;
-                            }
-
-                            break;
-                    }
-                }
-
-                return true;
-
-                static Binder getEnclosingBinderForNode(CSharpSyntaxNode contextNode, Binder contextBinder, SyntaxNode targetNode)
-                {
-                    while (true)
-                    {
-                        Binder? enclosingBinder = contextBinder.GetBinder(targetNode);
-
-                        if (enclosingBinder is not null)
-                        {
-                            return enclosingBinder;
-                        }
-
-                        if (targetNode == contextNode)
-                        {
-                            return contextBinder;
-                        }
-
-                        Debug.Assert(targetNode.Parent is not null);
-                        targetNode = targetNode.Parent;
-                    }
-                }
-
-                static bool childrenNeedChecking(SyntaxNode n)
-                {
-                    switch (n)
-                    {
-                        case MemberBindingExpressionSyntax:
-                        case BaseExpressionColonSyntax:
-                        case NameEqualsSyntax:
-                        case GotoStatementSyntax { RawKind: (int)SyntaxKind.GotoStatement }:
-                        case TypeParameterConstraintClauseSyntax:
-                        case AliasQualifiedNameSyntax:
-                            // These nodes do not have anything interesting for us
-                            return false;
-
-                        case AttributeListSyntax:
-                            // References in attributes, if any, are errors
-                            // Skip them
-                            return false;
-
-                        case ParameterSyntax:
-                            // Same as attributes
-                            return false;
-
-                        case AnonymousFunctionExpressionSyntax:
-                        case QueryExpressionSyntax:
-                            // Lambdas need special handling
-                            return false;
-
-                        case ExpressionSyntax expression:
-                            if (SyntaxFacts.IsInTypeOnlyContext(expression) &&
-                                !(expression.Parent is BinaryExpressionSyntax { RawKind: (int)SyntaxKind.IsExpression } isExpression &&
-                                    isExpression.Right == expression))
-                            {
-                                return false;
-                            }
-                            break;
-                    }
-
-                    return true;
-                }
-            }
-
-            protected abstract bool IsIdentifierOfInterest(IdentifierNameSyntax id);
-
-            private bool CheckLambda(AnonymousFunctionExpressionSyntax lambdaSyntax, Binder enclosingBinder)
-            {
-                UnboundLambda unboundLambda = enclosingBinder.AnalyzeAnonymousFunction(lambdaSyntax, BindingDiagnosticBag.Discarded);
-                var lambdaBodyBinder = CreateLambdaBodyBinder(enclosingBinder, unboundLambda);
-                return CheckIdentifiersInNode(lambdaSyntax.Body, lambdaBodyBinder.GetBinder(lambdaSyntax.Body) ?? lambdaBodyBinder);
-            }
-
-            private static ExecutableCodeBinder CreateLambdaBodyBinder(Binder enclosingBinder, UnboundLambda unboundLambda)
-            {
-                unboundLambda.HasExplicitReturnType(out RefKind refKind, out TypeWithAnnotations returnType);
-                var lambdaSymbol = new LambdaSymbol(
-                                        enclosingBinder,
-                                        enclosingBinder.Compilation,
-                                        enclosingBinder.ContainingMemberOrLambda!,
-                                        unboundLambda,
-                                        ImmutableArray<TypeWithAnnotations>.Empty,
-                                        ImmutableArray<RefKind>.Empty,
-                                        refKind,
-                                        returnType);
-
-                return new ExecutableCodeBinder(unboundLambda.Syntax, lambdaSymbol, unboundLambda.GetWithParametersBinder(lambdaSymbol, enclosingBinder));
-            }
-
-            private bool CheckIdentifier(Binder enclosingBinder, IdentifierNameSyntax id)
-            {
-                Debug.Assert(_lookupResult?.IsClear != false);
-                _lookupResult ??= LookupResult.GetInstance();
-
-                var useSiteInfo = CompoundUseSiteInfo<AssemblySymbol>.Discarded;
-                enclosingBinder.LookupIdentifier(_lookupResult, id, SyntaxFacts.IsInvoked(id), ref useSiteInfo);
-
-                return CheckAndClearLookupResult(enclosingBinder, id, _lookupResult);
-            }
-
-            protected abstract bool CheckAndClearLookupResult(Binder enclosingBinder, IdentifierNameSyntax id, LookupResult lookupResult);
-
-            protected static bool isTypeOrValueReceiver(
-                Binder enclosingBinder,
-                IdentifierNameSyntax id,
-                TypeSymbol type,
-                [NotNullWhen(true)] out SyntaxNode? memberAccessNode,
-                [NotNullWhen(true)] out string? memberName,
-                out int targetMemberArity,
-                out bool invoked)
-            {
-                memberAccessNode = null;
-                memberName = null;
-                targetMemberArity = 0;
-                invoked = false;
-
-                switch (id.Parent)
-                {
-                    case MemberAccessExpressionSyntax { RawKind: (int)SyntaxKind.SimpleMemberAccessExpression } memberAccess when memberAccess.Expression == id:
-                        var simpleName = memberAccess.Name;
-                        memberAccessNode = simpleName;
-                        memberName = simpleName.Identifier.ValueText;
-                        targetMemberArity = simpleName.Arity;
-                        invoked = SyntaxFacts.IsInvoked(memberAccess);
-                        break;
-                    case QualifiedNameSyntax qualifiedName when qualifiedName.Left == id:
-                        simpleName = qualifiedName.Right;
-                        memberAccessNode = simpleName;
-                        memberName = simpleName.Identifier.ValueText;
-                        targetMemberArity = simpleName.Arity;
-                        invoked = false;
-                        break;
-                    case FromClauseSyntax { Parent: QueryExpressionSyntax query } fromClause when query.FromClause == fromClause && fromClause.Expression == id:
-                        memberName = GetFirstInvokedMethodName(query, out memberAccessNode);
-                        targetMemberArity = 0;
-                        invoked = true;
-                        break;
-                }
-
-                return memberAccessNode is not null && enclosingBinder.IsPotentialColorColorReceiver(id, type);
-            }
-
-            /// <summary>
-            /// Follows the logic of <see cref="Binder.BindInstanceMemberAccess"/> and <see cref="Binder.BindMemberOfType"/>
-            /// </summary>
-            protected static bool TreatAsInstanceMemberAccess(
-                Binder enclosingBinder,
-                TypeSymbol type,
-                SyntaxNode memberAccessNode,
-                string memberName,
-                int targetMemberArity,
-                bool invoked,
-                LookupResult lookupResult)
-            {
-                Debug.Assert(!type.IsDynamic());
-                Debug.Assert(lookupResult.IsClear);
-
-                var useSiteInfo = CompoundUseSiteInfo<AssemblySymbol>.Discarded;
-                enclosingBinder.LookupInstanceMember(lookupResult, type, leftIsBaseReference: false, memberName, targetMemberArity, invoked, ref useSiteInfo);
-
-                bool treatAsInstanceMemberAccess;
-                if (lookupResult.IsMultiViable)
-                {
-                    // This branch follows the logic of BindMemberOfType
-                    Debug.Assert(lookupResult.Symbols.Any());
-
-                    var members = ArrayBuilder<Symbol>.GetInstance();
-                    Symbol symbol = enclosingBinder.GetSymbolOrMethodOrPropertyGroup(lookupResult, memberAccessNode, memberName, targetMemberArity, members, BindingDiagnosticBag.Discarded, wasError: out _,
-                                                                                     qualifierOpt: null);
-
-                    if ((object)symbol == null)
-                    {
-                        Debug.Assert(members.Count > 0);
-
-                        bool haveInstanceCandidates;
-                        lookupResult.Clear();
-                        enclosingBinder.CheckWhatCandidatesWeHave(members, type, memberName, targetMemberArity,
-                                                                  ref lookupResult, ref useSiteInfo,
-                                                                  out haveInstanceCandidates, out _);
-
-                        treatAsInstanceMemberAccess = haveInstanceCandidates;
-                    }
-                    else
-                    {
-                        // methods are special because of extension methods.
-                        Debug.Assert(symbol.Kind != SymbolKind.Method);
-                        treatAsInstanceMemberAccess = !(symbol.IsStatic || symbol.Kind == SymbolKind.NamedType);
-                    }
-
-                    members.Free();
-                }
-                else
-                {
-                    // At this point this could only be an extension method access or an error
-                    treatAsInstanceMemberAccess = true;
-                }
-
-                lookupResult.Clear();
-                return treatAsInstanceMemberAccess;
-            }
-
-            /// <summary>
-            /// Follows the logic of <see cref="Binder.BindQuery"/>
-            /// </summary>
-            private bool CheckQuery(QueryExpressionSyntax query, Binder enclosingBinder)
-            {
-                if (CheckIdentifiersInNode(query.FromClause.Expression, enclosingBinder))
-                {
-                    (QueryTranslationState state, _) = enclosingBinder.MakeInitialQueryTranslationState(query, BindingDiagnosticBag.Discarded);
-
-                    bool result = BindQueryInternal(enclosingBinder, state);
-
-                    for (QueryContinuationSyntax? continuation = query.Body.Continuation; continuation != null && result; continuation = continuation.Body.Continuation)
-                    {
-                        // A query expression with a continuation
-                        //     from ... into x ...
-                        // is translated into
-                        //     from x in ( from ... ) ...
-                        enclosingBinder.PrepareQueryTranslationStateForContinuation(state, continuation, BindingDiagnosticBag.Discarded);
-                        result = BindQueryInternal(enclosingBinder, state);
-                    }
-
-                    state.Free();
-                    return result;
-                }
-
-                return false;
-            }
-
-            /// <summary>
-            /// Follows the logic of <see cref="Binder.BindQueryInternal2"/>
-            /// </summary>
-            private bool BindQueryInternal(Binder enclosingBinder, QueryTranslationState state)
-            {
-                // we continue reducing the query until it is reduced away.
-                do
-                {
-                    if (state.clauses.IsEmpty())
-                    {
-                        return FinalTranslation(enclosingBinder, state);
-                    }
-                }
-                while (ReduceQuery(enclosingBinder, state));
-
-                return false;
-            }
-
-            /// <summary>
-            /// Follows the logic of <see cref="Binder.FinalTranslation"/>
-            /// </summary>
-            private bool FinalTranslation(Binder enclosingBinder, QueryTranslationState state)
-            {
-                Debug.Assert(state.clauses.IsEmpty());
-                switch (state.selectOrGroup.Kind())
-                {
-                    case SyntaxKind.SelectClause:
-                        {
-                            // A query expression of the form
-                            //     from x in e select v
-                            // is translated into
-                            //     ( e ) . Select ( x => v )
-                            var selectClause = (SelectClauseSyntax)state.selectOrGroup;
-                            var x = state.rangeVariable;
-                            var v = selectClause.Expression;
-                            return MakeQueryUnboundLambda(enclosingBinder, state.RangeVariableMap(), x, v);
-                        }
-                    case SyntaxKind.GroupClause:
-                        {
-                            // A query expression of the form
-                            //     from x in e group v by k
-                            // is translated into
-                            //     ( e ) . GroupBy ( x => k , x => v )
-                            // except when v is the identifier x, the translation is
-                            //     ( e ) . GroupBy ( x => k )
-                            var groupClause = (GroupClauseSyntax)state.selectOrGroup;
-                            var x = state.rangeVariable;
-                            var v = groupClause.GroupExpression;
-                            var k = groupClause.ByExpression;
-
-                            return MakeQueryUnboundLambda(enclosingBinder, state.RangeVariableMap(), x, k) &&
-                                   MakeQueryUnboundLambda(enclosingBinder, state.RangeVariableMap(), x, v);
-                        }
-                    default:
-                        {
-                            // there should have been a syntax error if we get here.
-                            return true;
-                        }
-                }
-            }
-
-            /// <summary>
-            /// Follows the logic of <see cref="Binder.ReduceQuery"/>
-            /// </summary>
-            private bool ReduceQuery(Binder enclosingBinder, QueryTranslationState state)
-            {
-                var topClause = state.clauses.Pop();
-                switch (topClause.Kind())
-                {
-                    case SyntaxKind.WhereClause:
-                        return ReduceWhere(enclosingBinder, (WhereClauseSyntax)topClause, state);
-                    case SyntaxKind.JoinClause:
-                        return ReduceJoin(enclosingBinder, (JoinClauseSyntax)topClause, state);
-                    case SyntaxKind.OrderByClause:
-                        return ReduceOrderBy(enclosingBinder, (OrderByClauseSyntax)topClause, state);
-                    case SyntaxKind.FromClause:
-                        return ReduceFrom(enclosingBinder, (FromClauseSyntax)topClause, state);
-                    case SyntaxKind.LetClause:
-                        return ReduceLet(enclosingBinder, (LetClauseSyntax)topClause, state);
-                    default:
-                        throw ExceptionUtilities.UnexpectedValue(topClause.Kind());
-                }
-            }
-
-            /// <summary>
-            /// Follows the logic of <see cref="Binder.ReduceWhere"/>
-            /// </summary>
-            private bool ReduceWhere(Binder enclosingBinder, WhereClauseSyntax where, QueryTranslationState state)
-            {
-                // A query expression with a where clause
-                //     from x in e
-                //     where f
-                //     ...
-                // is translated into
-                //     from x in ( e ) . Where ( x => f )
-                return MakeQueryUnboundLambda(enclosingBinder, state.RangeVariableMap(), state.rangeVariable, where.Condition);
-            }
-
-            /// <summary>
-            /// Follows the logic of <see cref="Binder.ReduceJoin"/>
-            /// </summary>
-            private bool ReduceJoin(Binder enclosingBinder, JoinClauseSyntax join, QueryTranslationState state)
-            {
-                if (CheckIdentifiersInNode(join.InExpression, enclosingBinder) &&
-                    MakeQueryUnboundLambda(enclosingBinder, state.RangeVariableMap(), state.rangeVariable, join.LeftExpression))
-                {
-                    var x2 = state.AddRangeVariable(enclosingBinder, join.Identifier, BindingDiagnosticBag.Discarded);
-
-                    if (MakeQueryUnboundLambda(enclosingBinder, QueryTranslationState.RangeVariableMap(x2), x2, join.RightExpression))
-                    {
-                        if (join.Into != null)
-                        {
-                            state.allRangeVariables[x2].Free();
-                            state.allRangeVariables.Remove(x2);
-
-                            state.AddRangeVariable(enclosingBinder, join.Into.Identifier, BindingDiagnosticBag.Discarded);
-                        }
-
-                        return true;
-                    }
-                }
-
-                return false;
-            }
-
-            /// <summary>
-            /// Follows the logic of <see cref="Binder.ReduceOrderBy"/>
-            /// </summary>
-            private bool ReduceOrderBy(Binder enclosingBinder, OrderByClauseSyntax orderby, QueryTranslationState state)
-            {
-                // A query expression with an orderby clause
-                //     from x in e
-                //     orderby k1 , k2 , ... , kn
-                //     ...
-                // is translated into
-                //     from x in ( e ) . 
-                //     OrderBy ( x => k1 ) . 
-                //     ThenBy ( x => k2 ) .
-                //     ... .
-                //     ThenBy ( x => kn )
-                //     ...
-                // If an ordering clause specifies a descending direction indicator,
-                // an invocation of OrderByDescending or ThenByDescending is produced instead.
-                foreach (var ordering in orderby.Orderings)
-                {
-                    if (!MakeQueryUnboundLambda(enclosingBinder, state.RangeVariableMap(), state.rangeVariable, ordering.Expression))
-                    {
-                        return false;
-                    }
-                }
-
-                return true;
-            }
-
-            /// <summary>
-            /// Follows the logic of <see cref="Binder.ReduceFrom"/>
-            /// </summary>
-            private bool ReduceFrom(Binder enclosingBinder, FromClauseSyntax from, QueryTranslationState state)
-            {
-                var x1 = state.rangeVariable;
-                if (MakeQueryUnboundLambda(enclosingBinder, state.RangeVariableMap(), x1, from.Expression))
-                {
-                    state.AddRangeVariable(enclosingBinder, from.Identifier, BindingDiagnosticBag.Discarded);
-                    return true;
-                }
-
-                return false;
-            }
-
-            /// <summary>
-            /// Follows the logic of <see cref="Binder.ReduceLet"/>
-            /// </summary>
-            private bool ReduceLet(Binder enclosingBinder, LetClauseSyntax let, QueryTranslationState state)
-            {
-                // A query expression with a let clause
-                //     from x in e
-                //     let y = f
-                //     ...
-                // is translated into
-                //     from * in ( e ) . Select ( x => new { x , y = f } )
-                //     ...
-                var x = state.rangeVariable;
-
-                if (MakeQueryUnboundLambda(enclosingBinder, state.RangeVariableMap(), x, let.Expression))
-                {
-                    state.rangeVariable = state.TransparentRangeVariable(enclosingBinder);
-                    state.AddTransparentIdentifier(x.Name);
-                    var y = state.AddRangeVariable(enclosingBinder, let.Identifier, BindingDiagnosticBag.Discarded);
-                    state.allRangeVariables[y].Add(let.Identifier.ValueText);
-                    return true;
-                }
-
-                return false;
-            }
-
-            private bool MakeQueryUnboundLambda(Binder enclosingBinder, RangeVariableMap qvm, RangeVariableSymbol parameter, ExpressionSyntax expression)
-            {
-                UnboundLambda unboundLambda = Binder.MakeQueryUnboundLambda(
-                    expression,
-                    new QueryUnboundLambdaState(
-                        enclosingBinder, qvm, ImmutableArray.Create(parameter),
-                        (LambdaSymbol lambdaSymbol, Binder lambdaBodyBinder, BindingDiagnosticBag diagnostics) => throw ExceptionUtilities.Unreachable()),
-                    withDependencies: false);
-
-                var lambdaBodyBinder = CreateLambdaBodyBinder(enclosingBinder, unboundLambda);
-                return CheckIdentifiersInNode(expression, lambdaBodyBinder.GetRequiredBinder(expression));
-            }
         }
     }
 }
